Move to using a sample type This is a custom type wrapper for []float32, with convenience methods defined on it for common sample operations, such as adding two samples, scaling by a constant gain, multiplying two samples, etc.
Moving to using this type enabled some nice refactoring in other
parts of the codebase. Less worrying about what's happening on each
channel of a signal, and more focus on what the actual behaviour
should be.

Also refactor some of the components in router, which were previously
returning a function that composed other components then called it,
but which now simply compose those components ahead of time and
return only the composed function for runtime evaluation.